Colors as floats until coercion to output scale
I don't know if this is a *good* way to tackle this problem, but I was
having all sorts of issues with integer overflow and thresholding when
they were 8-bit triples, so let's keep 'em flexible until I need to put
them on an actual pixel.

It does "work" in that things no longer look terrible.