Extract Hittable interface to replace Object
I want to stop returning references to hit objects, and boil down the
information returned when detecting hits, as a preface to extracting a
Material type that can handle scattering.

This extracts a new Hittable trait which our objects now implement, with
a hit() method that returns a Hit record. Currently the Hit record has
material-ish information living directly on it, including color and
reflectance; this will be extracted later.

I'm certain there's a more elegant way of handling the "find the closest
hit" question, but it is what it is for the time being.