Lambertian reflective material + skybox
This is the alternative approach to matte surfaces, where we scatter on
the surface and just reflect the color we see, allowing the scatter +
sampling to provide the matte effect.

I've only applied it to one thing in the scene so it looks all odd
because some things are lit by lights, while others are lit by the
skybox.