Don't scatter fuzzy reflections under the surface
If the fuzz vector takes the scattered ray below the surface of a
sphere, you get weird boundary effects. Ray Tracing in a Weekend solves
this by absorbing the reflected ray, but I don't have that in my
interface yet (I handle absorption with attenuation), so instead let's
loop til we find a valid scatter ray.