Fix fuzzy scatter infinite loop
Very occasionally, sphere intersections would be detected for rays that
point slightly in the same direction as the sphere's surface normal
(thanks to FP rounding).

Mostly this wasn't a problem, but for fuzzy reflections where we were
looping until the reflection vector points away from the surface, it
risked an infinite loop.

Rather than add more FP hacks, I've decided to follow "in a weekend"'s
approach, and simply absorb fuzzy scatters if the scatter isn't away
from the sphere. I don't intend to use this absorption model in general
(I'd prefer to trace and attenuate), but for this purpose I think it's
the simplest approach.