Eliminate colour clamping
So currrrrrently we don't have an opportunity for color value overflow,
so I'm removing the built-in clamping from the Color module. I think
we'll still need this at some point, but at the moment it's just
stinking up the joint so let's clean things up until we need it, and
know where we need it.