Get rid of std::vec::Vec allocation in the intersection calc
This code is the hot path; much as I like the filtery stuff it's
inefficient as hell; for a 100-sample render this cuts the time from 50s
to 25s.

Using a static(?) loop and a range inclusion is about as readable,
without the performance hit of the allocations caused by filtering on
iterators.

Technically I've altered the behaviour here because I check for
t_min <= t < t_max (as opposed to including t_max), but I don't think
this is important for float comparisons, so it looks odder to have an
inclusive range than not.