Don't fill-in rays that are absorbed or run out of depth
Currently the Option around the trace color return value is causing
absorbed or out-of-depth rays to bubble up and get assigned a grey
color. This is wrong; they should be black.

In fact we can completely get rid of the Option, because the skybox
color is now handled inside the trace() call, so we can guarantee a
color will come back.