Minimise random() calls
Introducing depth-of-field slowed down rendering considerably, the main
change being the randomisation of the ray start point.

Profiling shows a bunch of time spend in dynamic lib interactions during
random number generation. Since you can ask for as many random floats at
once as you like (I now know!), this allows us to minimise the number of
calls.

On a 1000-sample render this gave a 6.5min -> 4.5min speedup.

I'm also going to switch from the default, cryptographically-secure RNG
to a faster PRNG, but I'm struggling to get the newly-split-up rand
crates to play nicely together, and this is a good improvement anyway.