Please, my Monte Carlo, it's very secure
So, apparently the default thread_rng() provided by the rand crate is a
cryptographically secure one, so while it's still fast, it prioritises
security over performance.

I can therefore use a cheaper one, if I accept that in certain
circumstances, if someone stares very hard at one of my fuzzy spheres,
they'll be able to infer the state of my PRNG.

I've decided I can live with that.

There's an almost comically large diff here because:

1. I want to initialise an RNG per thread and pass it around
2. I got a bunch of weird type errors when I tried to do so
3. I don't know what I'm doing

Anyway this is what happens when you mash your code against the type
checker for long enough that it gives in and lets you run it.