Add bounding boxes to hittables
This is the first step for implementing Bounding Volume Hierarchies,
which let us speed up intersection considerably by making a series of
cheaper ray tests to cut down the search space before calculating
intersections with more complex objects.

This implements axis-aligned bounding boxes (AABBs in the books I
think?), and a pre-optimised method for detecting rays that pass through
them. Unlike Hittable::hit(), the BoundingBox::hit() method only needs
to return a boolean; if the box returns a hit we'll then proceed to
query the objects within it for the actual intersection.

Note that not all objects have a bounding box (infinite planes being the
example here), so the Hittable::bounding_box() method has an Option for
its return type.

The hit() method itself is effectively saying:

    tmin, tmax = 0., Infinity
    for each axis {
      calculate the ray entry and exit point in the axis
      swap the entry and exit point if the ray direction is negative
      update tmin and tmax using the entry and exit point
      if tmax <= tmin {
        one or more axes have incompatible ranges => no hit
      }
    }
    all ranges compatible => hit

There's an extensive discussion of why this is good in Raytracing - the
next week.

http://www.realtimerendering.com/raytracing/Ray%20Tracing_%20The%20Next%20Week.pdf