Introduce BoundedHittable in place of optional return
I started to implement bounded volume hierarchies, and immediately ran
into trouble on construction, because I needed to handle the case where
a hittable was passed that didn't have a bounding box.

Instead of detecting a None bounding box at runtime, though, we can use
the type system, and have a new type, BoundedHittable. BVH can then
insist only objects of this type be passed to it, which is p cool.