Add Bounded Volume Hierarchy
This adds a BVH type that lets us model the bounded parts of our scene
as a tree of objects, where leaves are only tested for intersections if
a ray passes through their parent node.

Note that this only results in efficiency if the tree is constructed so
as to divide up the objects in a manner that divides the scene space
well, and this commit makes no attempt to do so; I'm just getting to the
point where I can construct a BVH with two things in it, and treat it
just like any other hittable.