Well-constructed BVH with any bounded objects
Previously I added a trivial BVH that allowed for exactly two hittables.
This now adds a (hopefully) well-constructed BVH given arbitrary
hittables.

Following the in-a-weekend guy's notes, I'm taking in a hittable array,
sorting according to a random (but deterministic!) axis, splitting in
half and recursing on the sub-arrays.

I've not done any proper benchmarking on this but it looks a solid
improvement.

The code itself is ugly as heck, but I had real trouble making this look
pretty and also satisfying the borrow checker, so here it is. It's
probably inefficient too, but it only runs once at the start of each
render, so that's probably fine.

I turned BoundedHittable into a combo of two traits to avoid having to
repeat the Send + Sync constraint. Who knows if this is actually good.
Not me.