Extract object color to ConstantTexture
I'd like physical objects to support textures for varying color across
their surfaces.

I tried to implement this by adding texture to Material (since that's
how in-a-weekend does it), but ran into varying problems:

* Using a texture trait object caused lifetimes to end up in even more
  places
* Using generics for Material<T: Texture> similarly propagated
  everywhere, because Hit holds a reference to Material, and everything
  knows about Hit

I've therefore settled on what's actually the minimal change to what I
had; the physical objects are now generic over their texture type, and
they're responsible for calling up the color value at a given point on
their surface.

I think this is actually fine, and makes every bit as much sense as
Material doing this, if not more.

There's still the old bodge in place for the plane checkerboard, this'll
be extracted to a texture class next.