Remove last traces of reflectance
This is a holdover from when I was handling reflection and diffusion
using separate models. Now I appear to be all-in on the in-a-weekend
style monte carlo approach, everything's done with surface scattering
and there is no non-recursive "diffuse" hack.

(This was already effectively disabled by all my objects having 1.0
reflectance and the diffuse code having been removed.)