Pass whole Hit to Material::scatter()
I want to make Material responsible for handling textures, which will
require it knowing about u,v coords.

The natural place to store these is on the Hit record, so as a
preparatory step, rather than passing bits and bobs from the hit to
scatter(), just pass the whole Hit.