Move texture responsibilities onto Material
This undoes a decision I made earlier, which was to have the hittable
manage the coordination of hits, scattering (via materials) and
attenuation (via textures).

The following reasoning led me here:

1. Hit detection and scattering/attenuation need to be separate phases,
   because the latter is expensive, and only needs to be performed for
   the closest hit for a ray.
2. This means hit detection and scattering/attenuation need to be
   separate calls
3. Because there are hittables (such as BVH) that have no texture or
   material, we can't have scattering/attenuation as part of the
   hittable interface
4. Moreover, because there are hittables (such as BVH!) that return hit
   records on behalf of other objects, it's not even possible for the
   camera to know which object was hit (unless it's returned in the hit
   record, which seems wonky).

It therefore makes most sense to have `Hittable::hit()` stuff the `Hit`
record with all the cheap information necessary to scatter and
attenuate, but leave the material/texture as a separate second phase,
responsible for turning a ray and texture coordinates into a scattered
ray and attenuation.

So, that's what this commit does. Innit.