Add CheckerboardTexture
This uses trig functions instead of the abs logic I previously had,
checkering in 3D.

I've used cosines instead of the sine in the book cos I've got a plane
at y = 0, and it gets horrible float artifacts because sin(0) = 0 so you
just get noise, really.

I'm using the hit point as in the in-a-weekend book but I think this
would actually be much better in 2D space; not completely sure how to
manage that for an infinite plane rn though. Will see if I can eliminate
the hit point once I get u,v coords working with other texture types.