Rescale all non-emissive colors to 1.0
So. With the exception of the color generated by the sky box, all of
these colors really only take effect as an albedo, i.e. an attenuation
factor for received light.

It therefore makes sense for their components to be in the range [0,
1.0], instead of having the completely arbitrary 8-bit max value be
their max range.

This therefore scales them all back to be in the proper range, and does
the scaling to an 8-bit value right at the end. I might at some point
sort this out with the type system, e.g. with an Albedo type that's
actually constrained to these ranges, but not right now.

Note that we can still have color values > 1, and indeed we will do as
soon as we add emissive objects aka lights. It's up to us how we then
handle that; the code as implemented will simply clamp the values, so
over-bright objects are white. But we retain enough information to, say,
implement some sort of HDR approach, and that's fine too.