Randomise pixel processing order
Dunno if this is a good idea, but essentially cos cheap pixels come
right at the start (sky pixels), the progress bar is basically bobbins.

This therefore randomises the pixel processing order, so we get a
reasonable sample of pixel duration from the start.

However! This may be really silly. Like, for locality of reference
reasons, potentially? I dunno. I don't *think* it's the case cos they're
all operating over the same scene, but if nearby pixels need to
reference similar data then it's possible this will slow it down.
However, I think that's relatively unlikely, so whatevs.