Generate uv coords and use for checkerboard
I need uv coords to use image textures. Before implementing those, this
switches the checkerboard texture to uv coords, and implements uv for
both spheres and planes.

In the process of this I've converted planes to be constructed with a uv
axis pair. This has exposed the fact that I've got a left-handed
coordinate system, which inverts the intuition for the direction of the
cross product.

I can't quite be bothered to fix this right now, but I will, yes I will.

I've added a UVTexture to help debug uv calculations for other objects,
since it was helpful here.

Note that the the uv for the plane doesn't go from 0 to 1, but repeats
infinitely - any texture needs to handle uvs > 1, and in doing so they
should just use the fractional portion.