Add image texture!
Okay, so on the face of it this should've been easy, but actually proved
slightly tricky due to the borrow checker. I think this was because:

1. I made ImageTexture hold a reference to a trait object (DynamicImage)
   so as to not constrain the types of image I could use[1]
2. This is the first time a hittable needs to hold a reference to
   anything
3. I think, but am not certain, that rust is making assumptions about
   the lifetimes of the various vecs of Box<Hittable> I have knocking
   around
4. I therefore needed to add quite a few lifetime specifiers to make it
   clear to the borrow checker that the image outlives the scene that
   borrows it

As I write this, I've almost immediately realised this might be silly
and I could just use type parameters on ImageTexture, so I'll try that
next and we'll see.

Other things: the v coordinate appeared upside down for the image, hence
the (1 - v) in the pixel calculation. I think that's probably fine - the
top left is (0, 0) in pixel coords, but maybe it makes sense for uv to
start from the bottom left? I dunno, when I've got more things I'm sure
it'll become clear.

[1] Even as I write that, I realise this is a case for type parameters,
though. I'll give that a go next.