Use sines for checkerboard texture
I switched to cosines cos of a weird boundary condition for the plane at
y = 0, but now I'm using uv coordinates we can switch to sines - this
means you start at a 0 so you get a full phase before the first switch,
instead of half a phase.

You can see the effect at the top of the sphere where previously you got
a small longitudinal segment to start with.