Slightly mystical Perlin noise
Okay. This appears to work, or at least produces something that looks
similar to the in-a-weekend output.

However! It has a number of puzzling properties, and because I blindly
ported the algorithm from the book, I don't fully understand what, if
anything, I got wrong. They are:

1. perlin_interp output ranges from -0.5 to 0.5

I'm not sure why this is, although given the ranges of the inputs I'm
not completely surprised. However, in-a-weekend makes no mention of it,
and just returns the accumulator without modification. I need the output
to be in the range 0.0 to 1.0, so I've added 0.5.

2. perlin_interp output actually ranges from -0.500005124 to 0.5000132
   or something like that

Having adjusted for the 0.5 offset, I still had visual artifacts caused
by occasionally emitting negative colors. Say what you like, negative
color seems bad. Color > 1.0 is slightly less bad, but still doesn't
seem very sensible for a non-emissive surface. I've therefore clamped
the output values.

All told, not super satisfying, but I'd like to move on and I'm less
fussed about getting Perlin right than I am other stuff.