Use perlin noise correctly
It turns out there was a mistake in the "next weekend" book, and it
wasn't the perlin implementation that was incorrect (it should return
values in the range -1 to 1, biased towards the center), but the use of
it in the texture class.

This applies the right corrections to the noise value to produce a
legitimate albedo figure, and it all looks closer to the book.