Separate Vector3 and Unit3
I'd like to model directions and normals as unit vectors, distinct from
other vectors in that they're necessarily of unit length.

I've also, perhaps misguidedly, decided that it shouldn't be possible to
take the dot product of two vectors; only a vector and a unit, or two
units, can be dotted together. I might unpick this later; it's at odds
with the PBR approach.

I think I'm also at odds in insisting that normals be unit vectors, but
again I'm just going to persist with this and see where I get.