Return Config object from CLI
This feels cleaner than returning something holding a reference to the
parsed CLI gubbins. The intention is to allow other ways of getting
config into the program (for example if I wasm-ify this and try to stick
it on a webpage).

Could simplify the CLI internals now to avoid the private struct, but am
hedging my bets.