Pacify cargo clippy
It doesn't like BVH or CLI for reasons best left unquestioned, and it
kinda helpfully pointed me towards Option::map which automates null
checks. Not sure it's obviously better than an explicit check but maybe
that's just because I think of map as an array thing. Now I want a
burrito.